♻ refactor: タッチ関連をusePointerに切り出し
