+++ 20260104-that-instrument-innocencer/src/App.tsx
@@ -1,46 +1,23 @@
-import { useState } from 'react';
+import { usePointer } from './hooks/usePointer';
 import { useSound } from './hooks/useSound';
 import './App.css';
 
 function App() {
-  /* =====================
-   * State（UIに影響する状態）
-   * ===================== */
-  const [isDown, setIsDown] = useState(false);
-  const [position, setPosition] = useState({ x: 0, y: 0 });
-
+  const pointer = usePointer();
   const sound = useSound();
 
   /* =====================
    * Pointer（入力イベント）
    * ===================== */
-    const handlePointerDown = async (e: React.PointerEvent) => {
+  const handlePointerDown = async (e: React.PointerEvent) => {
     await sound.startAudio();
     sound.playSound();
-
-    setIsDown(true);
-    updatePosition(e);
-  };
-
-  const handlePointerMove = (e: React.PointerEvent) => {
-    if (isDown) {
-      updatePosition(e);
-    }
+    pointer.handlers.handlePointerDown(e);
   };
 
   const handlePointerUp = () => {
-    setIsDown(false);
     sound.stopSound();
-  };
-
-  /* =====================
-   * Utility
-   * ===================== */
-  const updatePosition = (e: React.PointerEvent) => {
-    setPosition({
-      x: Math.round(e.clientX),
-      y: Math.round(e.clientY),
-    });
+    pointer.handlers.handlePointerUp();
   };
 
   /* =====================
@@ -50,21 +27,21 @@ function App() {
     <div
       className="app-container"
       onPointerDown={handlePointerDown}
-      onPointerMove={handlePointerMove}
+      onPointerMove={pointer.handlers.handlePointerMove}
       onPointerUp={handlePointerUp}
       onPointerLeave={handlePointerUp}
       style={{ touchAction: 'none' }}
     >
       <div className="info-display">
-        <h1>{isDown ? 'TOUCHING' : 'WAITING'}</h1>
-        <p>X: {position.x}</p>
-        <p>Y: {position.y}</p>
+        <h1>{pointer.isDown ? 'TOUCHING' : 'WAITING'}</h1>
+        <p>X: {pointer.position.x}</p>
+        <p>Y: {pointer.position.y}</p>
 
         <div
-          className={`pointer-dot ${isDown ? 'active' : ''}`}
+          className={`pointer-dot ${pointer.isDown ? 'active' : ''}`}
           style={{
-            left: position.x,
-            top: position.y,
+            left: pointer.position.x,
+            top: pointer.position.y,
           }}
         />
       </div>
